+++ src/entities/index.ts
@@ -25,4 +25,4 @@ export * from './utils';
 export * from './swap';
 export * from './swap/swaps/v2/auraBalSwaps';
 export * from './addLiquidityBoosted';
-export * from './removeLiquidityBoostedV3';
+export * from './removeLiquidityBoosted';
